move config
