+++ src/streamdiffusion/__init__.py
@@ -1,43 +1,28 @@
 from .pipeline import StreamDiffusion
 from .wrapper import StreamDiffusionWrapper
+from .config import load_config, save_config, create_wrapper_from_config
 
 # ControlNet support
 try:
     from .controlnet import (
         ControlNetPipeline,
-        create_controlnet_pipeline,
-        ControlNetConfig,
-        StreamDiffusionControlNetConfig,
-        load_config,
-        save_config,
-        create_wrapper_from_config,
-        load_config,
-        BasePreprocessor,
-        CannyPreprocessor,
-        DepthPreprocessor,
-        OpenPosePreprocessor,
-        LineartPreprocessor,
-        get_preprocessor,
     )
 
+    
     __all__ = [
         "StreamDiffusion",
-        "StreamDiffusionWrapper",
-        "ControlNetPipeline",
-        "create_controlnet_pipeline",
-        "ControlNetConfig",
-        "StreamDiffusionControlNetConfig",
+        "StreamDiffusionWrapper", 
         "load_config",
         "save_config",
         "create_wrapper_from_config",
-        "load_config",
-        "BasePreprocessor",
-        "CannyPreprocessor",
-        "DepthPreprocessor",
-        "OpenPosePreprocessor",
-        "LineartPreprocessor",
-        "get_preprocessor",
+        "ControlNetPipeline",
     ]
 except ImportError as e:
     print(f"Warning: ControlNet support not available: {e}")
-    __all__ = ["StreamDiffusion", "StreamDiffusionWrapper"]
+    __all__ = [
+        "StreamDiffusion", 
+        "StreamDiffusionWrapper",
+        "load_config",
+        "save_config", 
+        "create_wrapper_from_config",
+    ]
